+++ docs/Whats-New.md
@@ -43,6 +43,8 @@ New:
 - Weapon targeting filter (by Uranusian)
 - Burst-specific FLH's for TechnoTypes (by Starkku)
 - Burst delays for weapons (by Starkku)
+- Powerplant Enhancer (by secsome)
+- Unlimited Global / Local Variables (by secsome)
 
 Vanilla fixes:
 - Fixed laser drawing code to allow for thicker lasers in house color draw mode (by Kerbiter, ChrisLv_CN)